+++ lib/interaction/src/manips/SoDirLtManip.cpp
@@ -507,7 +507,7 @@ SoDirectionalLightManip::getBoundingBox( SoGetBoundingBoxAction *action )
 
     // Now, set the center to be the average:
     if (numCenters != 0)
-	action->setCenter(totalCenter / numCenters, FALSE);
+	action->setCenter(totalCenter / (float)numCenters, FALSE);
 }
 
 void
